+++ CHANGES.rst
@@ -4,6 +4,8 @@ Change log
 1.0 (unreleased)
 ----------------
 
+- Added support for the SAML 2.0 Single Logout Service
+
 - Removed support for Python 3.7 and 3.8:
   To avoid compatibility issues between ``pysaml2`` and ``xmlsec1`` versions
   this package now requires at least Python 3.9. The ``pysaml2`` package loaded